requirements.txt
opencv_python==4.0.0.21
numpy==1.14.5
openpyxl==2.6.0
Pillow==6.0.0
